requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-material-admin==1.8.6
django-social-share==2.2.1
django-taggit==3.0.0
django-tinymce==3.5.0
numpy==1.24.2
pandas==1.5.3
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
